Que
In a correct C++ program, each opening group symbol should have a matching closing group symbol. For example a possible correct order of grouping symbols is [{(){()}}], while an incorrect order is [{(){(})}]. In addition to the traditional symbols, the C++ version we are using has one more delimiting symbol, $, which is used to print value of an expression. We call this version EC++. For example, 
int main(){
//some code
$5*(6+4/(3+2))$
//some code
}
Note that the grouping symbol should be balanced between two delimiting symbols.