Question
Consider the following problem: Given an input sequence a0, · · · , an−1 of n positive numbers, output a sequence b0,··· ,bn−1 where for any i ε {0,··· ,n − 1}, bi is the rightmost number to the left of ai that is bigger than ai in the given sequence. If no number to the left of ai is bigger than ai, then bi is defined to be −1. Thus b0 is always −1 because there is no number to the left of a0. For instance, if n = 7 and the input sequence is 7, 2, 5, 3, 6, 9, 1, then the  output sequence is −1, 7, 7, 5, 7, −1, 9.